backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
# psycopg2-binary==2.9.9  # Solo necesario para PostgreSQL, comentado para SQLite
alembic>=1.14.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
openai>=1.58.0
pillow>=10.2.0
pytesseract>=0.3.10
twilio>=9.5.0
httpx>=0.28.1
python-multipart>=0.0.12

